backend/requirements.txt
Django>=5.0,<6.0
djangorestframework
django-cors-headers
django-filter
djangorestframework-simplejwt
django-redis
redis
psycopg2-binary
gunicorn
WeasyPrint
Pillow
bleach
drf-spectacular
requests
python-dotenv
factory-boy
